Add game Bulls and Cows
Game generates secret code with desired length (from 1 to 36) and desired number of possible symbols (from 1 to 36).

You should guess code. If symbol is presented in secret code and placed in the same position as in the secret code, then it is a bull. If a symbol is presented in secret code but placed not in the same position, then it is a cow.

You can win the game only if all symbols are bulls.